libcnb-package: Only expose functions/types publicly if they are actually used
In order to:
- make the implementation of `libcnb-package` easier to reason about/refactor
- reduce the number of APIs visible to consumers of `libcnb-package`, so it's
  easier for them to see which APIs they should be using

In the case of `calculate_dir_size`, I've moved it back to `libcnb-cargo`
(where it was originally), since it's only used there.

I've checked that none of these APIs are used by our release automation:
https://github.com/search?q=repo%3Aheroku%2Flanguages-github-actions+libcnb_package+language%3ARust&type=code&l=Rust
(And longer term we should switch to having that automation call
a new `libcnb-cargo` subcommand rather than use `libcnb-package`
directly.)

I've not added a changelog entry, since we document `libcnb-package` as
having no stability guarantees:
https://github.com/heroku/libcnb.rs/blob/main/libcnb-package/README.md
